+++ src/components/modals/DiscardChangesModal.tsx
@@ -1,8 +1,8 @@
 import type { PropsWithChildren } from 'react'
-import { SolidButton } from '../user-action/Button'
 import type { PropsForTranslation } from '@/localization/useTranslation'
 import { useTranslation } from '@/localization/useTranslation'
-import { Modal, type ModalProps } from '../layout-and-navigation/Overlay'
+import type { ConfirmModalProps } from '@/components/modals/ConfirmModal'
+import { ConfirmModal } from '@/components/modals/ConfirmModal'
 
 type DiscardChangesModalTranslation = {
   save: string,
@@ -29,7 +29,7 @@ const defaultDiscardChangesModalTranslation = {
   }
 }
 
-type DiscardChangesModalProps = ModalProps & {
+type DiscardChangesModalProps = Omit<ConfirmModalProps, 'onDecline' | 'onConfirm' | 'buttonOverwrites'> & {
   isShowingDecline?: boolean,
   requireAnswer?: boolean,
   onCancel: () => void,
@@ -48,26 +48,19 @@ export const DiscardChangesModal = ({
                                      }: PropsForTranslation<DiscardChangesModalTranslation, PropsWithChildren<DiscardChangesModalProps>>) => {
   const translation = useTranslation(defaultDiscardChangesModalTranslation, overwriteTranslation)
   return (
-    <Modal
+    <ConfirmModal
       headerProps={{
         ...headerProps,
         titleText: headerProps?.titleText ?? translation.title,
         descriptionText: headerProps?.descriptionText ?? translation.description,
       }}
+      onConfirm={onSave}
+      onCancel={onCancel}
+      onDecline={onDontSave}
+      buttonOverwrites={[{ text: translation.cancel }, { text: translation.dontSave }, { text: translation.save }]}
       {...modalProps}
     >
       {children}
-      <div className="row mt-3 gap-x-4 justify-end">
-        <SolidButton color="positive" onClick={onSave}>
-          {translation.save}
-        </SolidButton>
-        <SolidButton color="negative" onClick={onDontSave}>
-          {translation.dontSave}
-        </SolidButton>
-        <SolidButton autoFocus color="primary" onClick={onCancel}>
-          {translation.cancel}
-        </SolidButton>
-      </div>
-    </Modal>
+    </ConfirmModal>
   )
 }