+++ stories/layout/dialog/Dialog.stories.tsx
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/nextjs'
 import { useState } from 'react'
-import { SolidButton } from '../../src'
-import { Dialog } from '../../src'
+import { Dialog, SolidButton } from '../../../src'
 
 
 /**
@@ -13,19 +12,21 @@ const StackedDialog = () => {
   const [modal3, setModal3] = useState(false)
   return (
     <>
-      <SolidButton onClick={() => setModal1(true)}>Open Modal</SolidButton>
+      <div className="row items-center justify-center min-h-[400px]">
+        <SolidButton onClick={() => setModal1(true)}>Open Modal</SolidButton>
+      </div>
 
-      <Dialog titleText="Dialog 1" isOpen={modal1} className="w-[300px]">
+      <Dialog headerProps={{ titleText: 'Dialog 1' }} isOpen={modal1} className="w-[300px]">
         <SolidButton onClick={() => setModal2(true)}>Open 2</SolidButton>
         <SolidButton color="negative" onClick={() => setModal1(false)}>Close</SolidButton>
       </Dialog>
 
-      <Dialog titleText="Dialog 2" isOpen={modal2} className="w-[400px] bg-green-300 text-black">
+      <Dialog headerProps={{ titleText: 'Dialog 2' }} isOpen={modal2} className="w-[400px] bg-green-300 text-black">
         <SolidButton onClick={() => setModal3(true)}>Open 3</SolidButton>
         <SolidButton color="negative" onClick={() => setModal2(false)}>Close</SolidButton>
       </Dialog>
 
-      <Dialog titleText="Dialog 3" isOpen={modal3}>
+      <Dialog headerProps={{ titleText: 'Dialog 3' }} isOpen={modal3}>
         <SolidButton color="negative" onClick={() => setModal3(false)}>Close 3</SolidButton>
       </Dialog>
     </>
@@ -33,7 +34,7 @@ const StackedDialog = () => {
 }
 
 const meta: Meta = {
-  title: 'Layout',
+  title: 'Layout/Dialog',
   component: StackedDialog,
 }