+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2025-06-20
+
+### Added
+
+- `ConfirmModal`
+
+### Changed
+
+- `ConfirmDialog` to not include a cancel option -> use the `ConfirmModal` instead
+- Change dark mode colors in `globals.css`
+
 ## [0.1.0] - 2025-06-04
 
 ### Added